// deleghebowcl/src/app/model/loginOperatore.ts
/**
 * Deleghe
 * Api risorse per gestire le informazioni di login dell'operatore
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/*******************************************************************************
* Copyright Regione Piemonte - 2022
* SPDX-License-Identifier: EUPL-1.2
******************************************************************************/
export interface LoginOperatore {
  codice_fiscale_operatore?: string;
  cognome_operatore?: string;
  nome_operatore?: string;
  esito_ricerca?: boolean;
  asl_operatore?: string;
  token_lcce?: string;
  ip_address?: string;
}