// deleghebowcl/src/app/model/dichiarazione.ts
/**
 * Deleghe
 * Api risorse per gestire le deleghe e le dichiarazioni (deleghe minori)
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/*******************************************************************************
* Copyright Regione Piemonte - 2022
* SPDX-License-Identifier: EUPL-1.2
******************************************************************************/
import {Cittadino} from './cittadino';
import {DichiarazioneDettaglio} from './dichiarazioneDettaglio';
import {DichiarazioneModo} from './dichiarazioneModo';
import {DichiarazioneStato} from './dichiarazioneStato';
import {DichiarazioneTipo} from './dichiarazioneTipo';


export interface Dichiarazione {
  uuid?: string;
  compilatoreCF?: string;
  npratica?: number;
  tipo?: DichiarazioneTipo;
  stato?: DichiarazioneStato;
  modo?: DichiarazioneModo;
  data_inserimento?: Date;
  cittadino?: Cittadino;
  dettagli?: Array<DichiarazioneDettaglio>;
  numero_documento?: string;
  autorita?: string;
  data_inizio_tutela?: string;
  data_fine_tutela?: string;
  stato_servizio_delega?: string;
  provenienza?: string;
}